+++ samples/net/sockets/echo_client/src/echo-client.c
@@ -229,7 +229,7 @@ static int check_our_ipv6_sockets(int sock,
 }
 
 static void ipv6_event_handler(struct net_mgmt_event_callback *cb,
-			       uint32_t mgmt_event, struct net_if *iface)
+			       uint64_t mgmt_event, struct net_if *iface)
 {
 	static char addr_str[INET6_ADDRSTRLEN];
 
@@ -297,7 +297,7 @@ static void ipv6_event_handler(struct net_mgmt_event_callback *cb,
 }
 
 static void event_handler(struct net_mgmt_event_callback *cb,
-			  uint32_t mgmt_event, struct net_if *iface)
+			  uint64_t mgmt_event, struct net_if *iface)
 {
 	if ((mgmt_event & EVENT_MASK) != mgmt_event) {
 		return;
