+++ subsys/net/lib/sockets/sockets_net_mgmt.c
@@ -31,7 +31,7 @@ __net_socket struct net_mgmt_socket {
 	uintptr_t pid;
 
 	/* net_mgmt mask */
-	uint32_t mask;
+	uint64_t mask;
 
 	/* Message allocation timeout */
 	k_timeout_t alloc_timeout;
@@ -145,7 +145,7 @@ static ssize_t znet_mgmt_recvfrom(struct net_mgmt_socket *mgmt, void *buf,
 {
 	struct sockaddr_nm *nm_addr = (struct sockaddr_nm *)src_addr;
 	k_timeout_t timeout = mgmt->wait_timeout;
-	uint32_t raised_event = 0;
+	uint64_t raised_event = 0;
 	uint8_t *copy_to = buf;
 	struct net_mgmt_msghdr hdr;
 	struct net_if *iface;
